move `salsa::query_group` code into its own module
